+++ otp/twofactor/totp2fa/totp.go
@@ -4,11 +4,13 @@ package totp2fa
 
 import (
 	"bytes"
+	"context"
 	"fmt"
 	"image/png"
 	"io"
 	"net/http"
 	"net/url"
+	"path"
 
 	"github.com/pkg/errors"
 	"github.com/pquerna/otp"
@@ -66,22 +68,47 @@ type TOTP struct {
 
 // Setup the module
 func (t *TOTP) Setup() error {
-	middleware := authboss.MountedMiddleware(t.Authboss, true, t.Authboss.Config.Modules.RoutesRedirectOnUnauthed, true, false)
-	t.Authboss.Core.Router.Get("/2fa/totp/setup", middleware(t.Core.ErrorHandler.Wrap(t.GetSetup)))
-	t.Authboss.Core.Router.Post("/2fa/totp/setup", middleware(t.Core.ErrorHandler.Wrap(t.PostSetup)))
+	var unauthedResponse authboss.MWRespondOnFailure
+	if t.Config.Modules.ResponseOnUnauthed != 0 {
+		unauthedResponse = t.Config.Modules.ResponseOnUnauthed
+	} else if t.Config.Modules.RoutesRedirectOnUnauthed {
+		unauthedResponse = authboss.RespondRedirect
+	}
+	abmw := authboss.MountedMiddleware2(t.Authboss, true, authboss.RequireFullAuth, unauthedResponse)
 
-	t.Authboss.Core.Router.Get("/2fa/totp/qr", middleware(t.Core.ErrorHandler.Wrap(t.GetQRCode)))
+	var middleware, verified func(func(w http.ResponseWriter, r *http.Request) error) http.Handler
+	middleware = func(handler func(http.ResponseWriter, *http.Request) error) http.Handler {
+		return abmw(t.Core.ErrorHandler.Wrap(handler))
+	}
 
-	t.Authboss.Core.Router.Get("/2fa/totp/confirm", middleware(t.Core.ErrorHandler.Wrap(t.GetConfirm)))
-	t.Authboss.Core.Router.Post("/2fa/totp/confirm", middleware(t.Core.ErrorHandler.Wrap(t.PostConfirm)))
+	if t.Authboss.Config.Modules.TwoFactorEmailAuthRequired {
+		setupPath := path.Join(t.Authboss.Paths.Mount, "/2fa/totp/setup")
+		emailVerify, err := twofactor.SetupEmailVerify(t.Authboss, "totp", setupPath)
+		if err != nil {
+			return err
+		}
+		verified = func(handler func(http.ResponseWriter, *http.Request) error) http.Handler {
+			return abmw(emailVerify.Wrap(t.Core.ErrorHandler.Wrap(handler)))
+		}
+	} else {
+		verified = middleware
+	}
 
-	t.Authboss.Core.Router.Get("/2fa/totp/remove", middleware(t.Core.ErrorHandler.Wrap(t.GetRemove)))
-	t.Authboss.Core.Router.Post("/2fa/totp/remove", middleware(t.Core.ErrorHandler.Wrap(t.PostRemove)))
+	t.Authboss.Core.Router.Get("/2fa/totp/setup", verified(t.GetSetup))
+	t.Authboss.Core.Router.Post("/2fa/totp/setup", verified(t.PostSetup))
+
+	t.Authboss.Core.Router.Get("/2fa/totp/qr", verified(t.GetQRCode))
+
+	t.Authboss.Core.Router.Get("/2fa/totp/confirm", verified(t.GetConfirm))
+	t.Authboss.Core.Router.Post("/2fa/totp/confirm", verified(t.PostConfirm))
+
+	t.Authboss.Core.Router.Get("/2fa/totp/remove", middleware(t.GetRemove))
+	t.Authboss.Core.Router.Post("/2fa/totp/remove", middleware(t.PostRemove))
 
 	t.Authboss.Core.Router.Get("/2fa/totp/validate", t.Core.ErrorHandler.Wrap(t.GetValidate))
 	t.Authboss.Core.Router.Post("/2fa/totp/validate", t.Core.ErrorHandler.Wrap(t.PostValidate))
 
-	t.Authboss.Events.Before(authboss.EventAuth, t.BeforeAuth)
+	t.Authboss.Events.Before(authboss.EventAuthHijack, t.HijackAuth)
 
 	return t.Authboss.Core.ViewRenderer.Load(
 		PageTOTPSetup,
@@ -93,9 +120,9 @@ func (t *TOTP) Setup() error {
 	)
 }
 
-// BeforeAuth stores the user's pid in a special temporary session variable
+// HijackAuth stores the user's pid in a special temporary session variable
 // and redirects them to the validation endpoint.
-func (t *TOTP) BeforeAuth(w http.ResponseWriter, r *http.Request, handled bool) (bool, error) {
+func (t *TOTP) HijackAuth(w http.ResponseWriter, r *http.Request, handled bool) (bool, error) {
 	if handled {
 		return false, nil
 	}
@@ -318,6 +345,7 @@ func (t *TOTP) PostValidate(w http.ResponseWriter, r *http.Request) error {
 	case err != nil:
 		return err
 	case !ok:
+		r = r.WithContext(context.WithValue(r.Context(), authboss.CTXKeyUser, user))
 		handled, err := t.Authboss.Events.FireAfter(authboss.EventAuthFail, w, r)
 		if err != nil {
 			return err
@@ -341,6 +369,7 @@ func (t *TOTP) PostValidate(w http.ResponseWriter, r *http.Request) error {
 
 	logger.Infof("user %s totp 2fa success", user.GetPID())
 
+	r = r.WithContext(context.WithValue(r.Context(), authboss.CTXKeyUser, user))
 	handled, err := t.Authboss.Events.FireAfter(authboss.EventAuth, w, r)
 	if err != nil {
 		return err